chat-app/requirements.txt
streamlit
sqlalchemy
psycopg2-binary
python-dotenv
langchain
langchain-groq
werkzeug